requirements.txt
# Core dependencies
python-dotenv==1.0.0
sqlalchemy==2.0.23
pandas==2.1.3
numpy==1.26.2
streamlit==1.28.2
pyautogen==0.2.0  # AutoGen framework

# Database drivers and connectors
pyodbc==5.0.1
pymysql==1.1.0
psycopg2-binary==2.9.9
oracledb==1.4.2
cx_Oracle==8.3.0
alembic==1.12.1

# Visualization libraries
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0
bokeh==3.3.0
pydeck==0.8.0

# Report generation
weasyprint==60.1
openpyxl==3.1.2
xlsxwriter==3.1.9
pdfkit==1.0.0
jinja2==3.1.2

# Security and encryption
cryptography==41.0.5
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
keyring==24.3.0

# Caching and performance
redis==5.0.1
cachetools==5.3.2
diskcache==5.6.3
joblib==1.3.2

# Machine learning and statistics
scikit-learn==1.3.2
scipy==1.11.3
statsmodels==0.14.0
prophet==1.1.4

# Testing and development
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.0
isort==5.12.0
pytest-html==4.1.1
pytest-xdist==3.3.1  # For parallel test execution
pytest-timeout==2.2.0
coverage==7.3.2

# Type checking
types-python-dateutil==2.8.19.14
types-requests==2.31.0.10
types-PyMySQL==1.1.0.1
types-psycopg2==2.9.21.11
types-redis==4.6.0.7

# Monitoring and logging
prometheus-client==0.18.0
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
structlog==23.2.0

# API and networking
requests==2.31.0
aiohttp==3.9.0
httpx==0.25.1
urllib3==2.1.0
fastapi==0.104.1  # For API endpoints
uvicorn==0.24.0  # ASGI server

# Utilities
tqdm==4.66.1
python-dateutil==2.8.2
pytz==2023.3.post1
tenacity==8.2.3
humanize==4.8.0

# LLM and embeddings
openai==1.3.5
anthropic==0.5.0
tiktoken==0.5.1
sentence-transformers==2.2.2
torch==2.1.1
transformers==4.35.2
tokenizers==0.15.0
einops==0.7.0
accelerate==0.24.1
